+++ examples/minimum/locals.tf
@@ -3,11 +3,19 @@ locals {
   vpc_id                    = data.aws_vpc.supporting.id
   database_subnets          = local.database_subnet_id
   supporting_resources_name = "terraform-aws-rds"
+
   database_subnet_id = [
     for s in data.aws_subnet.database : s.id
   ]
+
   tags = {
-    Name        = local.db_name
-    Environment = "examples"
+    Environment        = "example"
+    Name               = local.db_name
+    "user::CostCenter" = "terraform-registry"
+    Department         = "DevOps"
+    Project            = "Examples"
+    Owner              = "Boldlink"
+    LayerName          = "cExample"
+    LayerId            = "cExample"
   }
 }