+++ examples/s3-import/data.tf
@@ -1,4 +1,3 @@
-###
 data "aws_partition" "current" {}
 
 data "aws_iam_policy_document" "monitoring" {
@@ -22,13 +21,32 @@ data "aws_kms_alias" "rds" {
 data "aws_iam_policy_document" "s3_bucket" {
   version = "2012-10-17"
   statement {
-    sid     = "AllowAccess"
-    effect  = "Allow"
-    actions = ["s3:*"]
-    resources = [
-      "arn:aws:s3:::${aws_s3_bucket.mysql.bucket}",
-      "arn:aws:s3:::${aws_s3_bucket.mysql.bucket}/*"
+    sid    = "AllowAccess"
+    effect = "Allow"
+    actions = [
+      "s3:PutObject*",
+      "s3:ListBucket",
+      "s3:GetObject*",
+      "s3:DeleteObject*",
+      "s3:GetBucketLocation"
     ]
+    resources = [
+      "arn:${local.partition}:s3:::${local.name}",
+      "arn:${local.partition}:s3:::${local.name}/*"
+    ]
+  }
+}
+
+data "aws_iam_policy_document" "assume_policy" {
+  statement {
+    actions = [
+      "sts:AssumeRole",
+    ]
+
+    principals {
+      type        = "Service"
+      identifiers = ["rds.amazonaws.com"]
+    }
   }
 }